[DeadCode] Move RecastingRemovalRector from PHPStan set
