+++ config/set/phpstan.php
@@ -2,8 +2,8 @@
 
 declare(strict_types=1);
 
+use Rector\DeadCode\Rector\Cast\RecastingRemovalRector;
 use Rector\PHPStan\Rector\Assign\PHPStormVarAnnotationRector;
-use Rector\PHPStan\Rector\Cast\RecastingRemovalRector;
 use Rector\PHPStan\Rector\Node\RemoveNonExistingVarAnnotationRector;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
