+++ rules/dead-code/tests/Rector/Cast/RecastingRemovalRector/Fixture/fixture2.php.inc
